Split out types and resolvers into files under types/ and resolvers/.
