+++ src/index.js
@@ -1,18 +1,10 @@
 require('dotenv').config();
 const { ApolloServer } = require('apollo-server');
 
-// dependencies
-const { GraphQLFileLoader } = require("@graphql-tools/graphql-file-loader");
-const { loadSchemaSync } = require("@graphql-tools/load");
+const typeDefs = require('./typeDefs.js');
+const resolvers = require('./resolvers.js');
 
-// local
-const { resolvers } = require('./resolvers.js');
 const { dataSources } = require('./data-sources');
-
-// load the GraphQL schema with GraphQTLFileLoader
-const typeDefs = loadSchemaSync('schema.graphql', {
-  loaders: [new GraphQLFileLoader()],
-});
 
 // create the Apollo server
 const server = new ApolloServer({
