+++ src/resolvers/Ontology.js
@@ -1,0 +1,11 @@
+module.exports = {
+    
+    Query: {
+
+        ontology: async (_source, { id }, { dataSources }) => {
+            return dataSources.lisIntermineAPI.getOntology(id);
+        },
+
+    }
+
+}
