+++ src/resolvers/ProteinDomain.js
@@ -1,0 +1,30 @@
+module.exports = {
+    
+    Query: {
+
+        proteinDomain: async (_source, { id }, { dataSources }) => {
+            return dataSources.lisIntermineAPI.getProteinDomain(id);
+        }, 
+
+    },
+
+    ProteinDomain: {
+        genes: async (proteinDomain, { start, size }, { dataSources }) => {
+            const args = {
+                proteinDomain: proteinDomain,
+                start,
+                size,
+            };
+            return dataSources.lisIntermineAPI.getGenes(args);
+        },
+        geneFamilies: async (proteinDomain, { start, size }, { dataSources }) => {
+            const args = {
+                proteinDomain: proteinDomain,
+                start,
+                size,
+            };
+            return dataSources.lisIntermineAPI.getGeneFamilies(args);
+        },
+    },
+
+}
