+++ src/resolvers/QTL.js
@@ -1,0 +1,11 @@
+module.exports = {
+    
+    Query: {
+
+        qtl:  async (_source, { id }, { dataSources }) => {
+            return dataSources.lisIntermineAPI.getQTL(id);
+        },
+
+    }
+
+}
