+++ src/resolvers/README.md
@@ -1,0 +1,4 @@
+***resolvers***
+
+Each GraphQL type's resolvers are defined in a Javascript file here, e.g. Gene.js.
+
