+++ src/typeDefs.js
@@ -1,0 +1,11 @@
+// ----------------------------------------------------------------
+// Load the typeDefs from all .graphql files in the types directory
+// ----------------------------------------------------------------
+
+const path = require('path');
+const { loadFilesSync } = require('@graphql-tools/load-files');
+const { mergeTypeDefs } = require('@graphql-tools/merge');
+
+const typesArray = loadFilesSync(path.join(__dirname, './types'));
+
+module.exports = mergeTypeDefs(typesArray);
