+++ src/types/GWAS.graphql
@@ -1,0 +1,24 @@
+# <class name="GWAS" extends="Annotatable" is-interface="true" term="">
+# 	<attribute name="genotypingPlatform" type="java.lang.String"/>
+# 	<attribute name="description" type="java.lang.String"/>
+# 	<attribute name="genotypes" type="java.lang.String"/>
+# 	<attribute name="genotypingMethod" type="java.lang.String"/>
+# 	<attribute name="synopsis" type="java.lang.String"/>
+# 	<reference name="organism" referenced-type="Organism"/>
+# 	<reference name="dataSet" referenced-type="DataSet"/>
+# 	<collection name="results" referenced-type="GWASResult" reverse-reference="gwas"/>
+# </class>
+type GWAS {
+  id: ID!
+  identifier: String!
+  # ontologyAnnotations
+  # publications
+  genotypingPlatform: String
+  description: String
+  genotypes: String
+  genotypingMethod: String
+  synopsis: String
+  organism: Organism
+  # dataSet
+  results: [GWASResult]
+}
