+++ src/types/GWASResult.graphql
@@ -1,0 +1,17 @@
+# <class name="GWASResult" is-interface="true" term="">
+# 	<attribute name="pValue" type="java.lang.Double"/>
+# 	<attribute name="markerName" type="java.lang.String"/>
+# 	<reference name="gwas" referenced-type="GWAS" reverse-reference="results"/>
+# 	<reference name="trait" referenced-type="Trait" reverse-reference="gwasResults"/>
+# 	<reference name="dataSet" referenced-type="DataSet"/>
+# 	<collection name="markers" referenced-type="GeneticMarker" reverse-reference="gwasResults"/>
+# </class>
+type GWASResult {
+  id: ID!
+  pValue: Float
+  markerName: String
+  gwas: GWAS
+  trait: Trait
+  # dataSet
+  # markers
+}
