+++ src/types/Gene.graphql
@@ -1,0 +1,61 @@
+# <class name="Gene" extends="SequenceFeature" is-interface="true" term="http://purl.obolibrary.org/obo/SO_0000704,http://purl.obolibrary.org/obo/SO:0000704">
+# 	<attribute name="briefDescription" type="java.lang.String" term="http://semanticscience.org/resource/SIO_000136"/>
+# 	<reference name="upstreamIntergenicRegion" referenced-type="IntergenicRegion"/>
+# 	<reference name="downstreamIntergenicRegion" referenced-type="IntergenicRegion"/>
+# 	<collection name="flankingRegions" referenced-type="GeneFlankingRegion" reverse-reference="gene"/>
+# 	<collection name="transcripts" referenced-type="Transcript" reverse-reference="gene"/>
+# 	<collection name="introns" referenced-type="Intron" reverse-reference="genes"/>
+# 	<collection name="proteins" referenced-type="Protein" reverse-reference="genes" term="https://semanticscience.org/resource/SIO_010078"/>
+# 	<collection name="alleles" referenced-type="Allele" reverse-reference="gene"/>
+# 	<collection name="geneFamilyAssignments" referenced-type="GeneFamilyAssignment"/>
+# 	<collection name="CDSs" referenced-type="CDS" reverse-reference="gene"/>
+# 	<collection name="proteinDomains" referenced-type="ProteinDomain" reverse-reference="genes"/>
+# 	<collection name="exons" referenced-type="Exon" reverse-reference="gene"/>
+# 	<collection name="pathways" referenced-type="Pathway" reverse-reference="genes"/>
+# 	<collection name="regulatoryRegions" referenced-type="RegulatoryRegion" reverse-reference="gene"/>
+# 	<collection name="UTRs" referenced-type="UTR" reverse-reference="gene"/>
+# </class>
+type Gene implements SequenceFeature {
+  id: ID!
+  identifier: String!
+  # ontologyAnnotations
+  # publications
+  description: String
+  symbol: String
+  name: String
+  assemblyVersion: String
+  annotationVersion: String
+  organism: Organism
+  strain: Strain
+  # locations
+  # synonyms
+  # crossReferences
+  # dataSets
+  # locatedFeatures
+  # score
+  # scoreType
+  length: Int
+  # sequenceOntologyTerm
+  # supercontigLocation
+  # chromosomeLocation
+  # supercontig
+  # sequence
+  # chromosome
+  # overlappingFeatures
+  # childFeatures
+  # briefDescription
+  # upstreamIntergenicRegion
+  # downstreamIntergenicRegion
+  # flankingRegions
+  # transcripts
+  # introns
+  # proteins
+  # alleles
+  geneFamilyAssignments: [GeneFamilyAssignment]
+  # CDSs
+  proteinDomains: [ProteinDomain]
+  # exons
+  # pathways
+  # regulatoryRegions
+  # UTRs
+}
