+++ src/types/GeneFamilyAssignment.graphql
@@ -1,0 +1,13 @@
+# <class name="GeneFamilyAssignment" is-interface="true" term="">
+# 	<attribute name="bestDomainScore" type="java.lang.Double"/>
+# 	<attribute name="score" type="java.lang.Double"/>
+# 	<attribute name="evalue" type="java.lang.Double"/>
+# 	<reference name="geneFamily" referenced-type="GeneFamily"/>
+# </class>
+type GeneFamilyAssignment {
+  id: ID!
+  bestDomainScore: Float
+  score:  Float
+  evalue: Float
+  geneFamily: GeneFamily!
+}
