+++ src/types/Ontology.graphql
@@ -1,0 +1,11 @@
+# <class name="Ontology" is-interface="true" term="http://semanticscience.org/resource/SIO_001391">
+# 	<attribute name="url" type="java.lang.String" term="http://edamontology.org/data_1052"/>
+# 	<attribute name="name" type="java.lang.String" term="http://www.w3.org/2000/01/rdf-schema#label"/>
+# 	<collection name="dataSets" referenced-type="DataSet"/>
+# </class>
+type Ontology {
+  id: ID!
+  url: String
+  name: String
+  # dataSets
+}
