+++ src/types/Phylotree.graphql
@@ -1,0 +1,16 @@
+# <class name="Phylotree" extends="Annotatable" is-interface="true" term="">
+# 	<attribute name="numLeaves" type="java.lang.Integer"/>
+# 	<reference name="geneFamily" referenced-type="GeneFamily" reverse-reference="phylotree"/>
+# 	<collection name="nodes" referenced-type="Phylonode" reverse-reference="tree"/>
+# 	<collection name="dataSets" referenced-type="DataSet"/>
+# </class>
+type Phylotree {
+  id: ID!
+  identifier: String!
+  # ontologyAnnotations
+  # publications
+  numLeaves: Int
+  geneFamily: GeneFamily
+  nodes: [Phylonode]
+  # dataSets
+}
