+++ src/types/Strain.graphql
@@ -1,0 +1,18 @@
+# <class name="Strain" is-interface="true" term="http://semanticscience.org/resource/SIO_010055">
+# 	<attribute name="identifier" type="java.lang.String" term="http://edamontology.org/data_2379"/>
+# 	<attribute name="name" type="java.lang.String" term="http://edamontology.org/data_1046"/>
+# 	<attribute name="description" type="java.lang.String"/>
+# 	<attribute name="origin" type="java.lang.String"/>
+# 	<attribute name="accession" type="java.lang.String" term="http://edamontology.org/data_2912"/>
+# 	<reference name="organism" referenced-type="Organism" reverse-reference="strains" term="http://purl.org/net/orth#organism"/>
+# 	<collection name="dataSets" referenced-type="DataSet"/>
+# </class>
+type Strain {
+  id: ID!
+  identifier: String!
+  name: String
+  description: String
+  origin: String
+  accession: String
+  organism: Organism
+}
