+++ src/types/interfaces.graphql
@@ -1,0 +1,92 @@
+# ------------------------------------------------------------------------------
+# These interfaces correspond to widely-extended objects in the mine data model.
+# ------------------------------------------------------------------------------
+
+# <class name="Annotatable" is-interface="true">
+# 	<attribute name="primaryIdentifier" type="java.lang.String" term="http://semanticscience.org/resource/SIO_000673"/>
+# 	<collection name="ontologyAnnotations" referenced-type="OntologyAnnotation" reverse-reference="subject" term="http://semanticscience.org/resource/SIO_000255"/>
+# 	<collection name="publications" referenced-type="Publication" reverse-reference="entities" term="http://purl.org/dc/terms/bibliographicCitation"/>
+# </class>
+interface Annotatable {
+  id: ID!
+  identifier: String!
+  # ontologyAnnotations
+  # publications
+}
+
+# <class name="BioEntity" extends="Annotatable" is-interface="true" term="">
+# 	<attribute name="description" type="java.lang.String"/>
+# 	<attribute name="symbol" type="java.lang.String" term="http://www.w3.org/2004/02/skos/core#prefLabel"/>
+# 	<attribute name="name" type="java.lang.String" term="http://www.w3.org/2000/01/rdf-schema#label"/>
+# 	<attribute name="assemblyVersion" type="java.lang.String"/>
+# 	<attribute name="annotationVersion" type="java.lang.String"/>
+# 	<attribute name="secondaryIdentifier" type="java.lang.String" term="http://semanticscience.org/resource/SIO_000673"/>
+# 	<reference name="organism" referenced-type="Organism" term="http://purl.org/net/orth#organism"/>
+# 	<reference name="strain" referenced-type="Strain" term="http://semanticscience.org/resource/SIO_010055"/>
+# 	<collection name="locations" referenced-type="Location" reverse-reference="feature"/>
+# 	<collection name="synonyms" referenced-type="Synonym" reverse-reference="subject" term="http://purl.obolibrary.org/obo/synonym"/>
+# 	<collection name="crossReferences" referenced-type="CrossReference" reverse-reference="subject" term="http://www.geneontology.org/formats/oboInOwl#hasDbXref"/>
+# 	<collection name="dataSets" referenced-type="DataSet" reverse-reference="bioEntities" term="http://semanticscience.org/resource/SIO_001278"/>
+# 	<collection name="locatedFeatures" referenced-type="Location" reverse-reference="locatedOn"/>
+# </class>
+interface BioEntity {
+  id: ID!
+  identifier: String!
+  # ontologyAnnotations
+  # publications
+  description: String
+  symbol: String
+  name: String
+  assemblyVersion: String
+  annotationVersion: String
+  organism: Organism
+  strain: Strain
+  # locations
+  # synonyms
+  # crossReferences
+  # dataSets
+  # locatedFeatures
+}
+
+# <class name="SequenceFeature" extends="BioEntity" is-interface="true" term="http://purl.obolibrary.org/obo/SO_0000110,http://purl.obolibrary.org/obo/SO:0000110">
+# 	<attribute name="score" type="java.lang.Double" term="http://edamontology.org/data_1772"/>
+# 	<attribute name="scoreType" type="java.lang.String" term="http://purl.org/dc/terms/type"/>
+# 	<attribute name="length" type="java.lang.Integer" term="http://purl.org/dc/terms/SizeOrDuration"/>
+# 	<reference name="sequenceOntologyTerm" referenced-type="SOTerm"/>
+# 	<reference name="supercontigLocation" referenced-type="Location"/>
+# 	<reference name="chromosomeLocation" referenced-type="Location"/>
+# 	<reference name="supercontig" referenced-type="Supercontig"/>
+# 	<reference name="sequence" referenced-type="Sequence"/>
+# 	<reference name="chromosome" referenced-type="Chromosome" term="http://purl.org/dc/terms/isPartOf"/>
+# 	<collection name="overlappingFeatures" referenced-type="SequenceFeature"/>
+# 	<collection name="childFeatures" referenced-type="SequenceFeature"/>
+# </class>
+interface SequenceFeature {
+  id: ID!
+  identifier: String!
+  # ontologyAnnotations
+  # publications
+  description: String
+  symbol: String
+  name: String
+  assemblyVersion: String
+  annotationVersion: String
+  organism: Organism
+  strain: Strain
+  # locations
+  # synonyms
+  # crossReferences
+  # dataSets
+  # locatedFeatures
+  # score
+  # scoreType
+  length: Int
+  # sequenceOntologyTerm
+  # supercontigLocation
+  # chromosomeLocation
+  # supercontig
+  # sequence
+  # chromosome
+  # overlappingFeatures
+  # childFeatures
+}
